move tck tests to the specific package
