+++ vertx-mysql-client/src/test/java/io/vertx/mysqlclient/tck/MySQLBinaryDataTypeDecodeTest.java
@@ -1,7 +1,7 @@
 package io.vertx.mysqlclient.tck;
 
 import io.vertx.mysqlclient.junit.MySQLRule;
-import io.vertx.sqlclient.BinaryDataTypeDecodeTestBase;
+import io.vertx.sqlclient.tck.BinaryDataTypeDecodeTestBase;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import io.vertx.sqlclient.Row;
