+++ vertx-mysql-client/src/test/java/io/vertx/mysqlclient/tck/MySQLBinaryDataTypeEncodeTest.java
@@ -3,7 +3,7 @@ package io.vertx.mysqlclient.tck;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import io.vertx.mysqlclient.junit.MySQLRule;
-import io.vertx.sqlclient.BinaryDataTypeEncodeTestBase;
+import io.vertx.sqlclient.tck.BinaryDataTypeEncodeTestBase;
 import io.vertx.sqlclient.Row;
 import io.vertx.sqlclient.Tuple;
 import org.junit.ClassRule;
