+++ vertx-mysql-client/src/test/java/io/vertx/mysqlclient/tck/MySQLConnectionTest.java
@@ -1,7 +1,7 @@
 package io.vertx.mysqlclient.tck;
 
 import io.vertx.mysqlclient.junit.MySQLRule;
-import io.vertx.sqlclient.ConnectionTestBase;
+import io.vertx.sqlclient.tck.ConnectionTestBase;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.junit.ClassRule;
