+++ vertx-mysql-client/src/test/java/io/vertx/mysqlclient/tck/MySQLPreparedQueryCachedTest.java
@@ -3,7 +3,7 @@ package io.vertx.mysqlclient.tck;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import io.vertx.mysqlclient.junit.MySQLRule;
-import io.vertx.sqlclient.PreparedQueryCachedTestBase;
+import io.vertx.sqlclient.tck.PreparedQueryCachedTestBase;
 import org.junit.ClassRule;
 import org.junit.Ignore;
 import org.junit.Test;
