+++ vertx-mysql-client/src/test/java/io/vertx/mysqlclient/tck/MySQLSimpleQueryTest.java
@@ -17,7 +17,7 @@
 package io.vertx.mysqlclient.tck;
 
 import io.vertx.mysqlclient.junit.MySQLRule;
-import io.vertx.sqlclient.SimpleQueryTestBase;
+import io.vertx.sqlclient.tck.SimpleQueryTestBase;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.junit.ClassRule;
 import org.junit.runner.RunWith;
