+++ vertx-pg-client/src/test/java/io/vertx/pgclient/tck/PgBinaryDataTypeDecodeTest.java
@@ -1,7 +1,7 @@
 package io.vertx.pgclient.tck;
 
 import io.vertx.pgclient.junit.PgRule;
-import io.vertx.sqlclient.BinaryDataTypeDecodeTestBase;
+import io.vertx.sqlclient.tck.BinaryDataTypeDecodeTestBase;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.junit.ClassRule;
 import org.junit.runner.RunWith;
