+++ vertx-pg-client/src/test/java/io/vertx/pgclient/tck/PgConnectionTest.java
@@ -1,7 +1,7 @@
 package io.vertx.pgclient.tck;
 
 import io.vertx.pgclient.junit.PgRule;
-import io.vertx.sqlclient.ConnectionTestBase;
+import io.vertx.sqlclient.tck.ConnectionTestBase;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.junit.ClassRule;
