+++ vertx-pg-client/src/test/java/io/vertx/pgclient/tck/PgPreparedQueryCachedTest.java
@@ -2,7 +2,7 @@ package io.vertx.pgclient.tck;
 
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import io.vertx.pgclient.junit.PgRule;
-import io.vertx.sqlclient.PreparedQueryCachedTestBase;
+import io.vertx.sqlclient.tck.PreparedQueryCachedTestBase;
 import org.junit.ClassRule;
 import org.junit.runner.RunWith;
