+++ vertx-pg-client/src/test/java/io/vertx/pgclient/tck/PgSimpleQueryPooledTest.java
@@ -17,7 +17,7 @@
 package io.vertx.pgclient.tck;
 
 import io.vertx.pgclient.junit.PgRule;
-import io.vertx.sqlclient.SimpleQueryTestBase;
+import io.vertx.sqlclient.tck.SimpleQueryTestBase;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.junit.ClassRule;
 import org.junit.runner.RunWith;
