+++ vertx-sql-client/src/test/java/io/vertx/sqlclient/tck/ConnectionTestBase.java
@@ -1,10 +1,12 @@
-package io.vertx.sqlclient;
+package io.vertx.sqlclient.tck;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
+import io.vertx.sqlclient.SqlConnectOptions;
+import io.vertx.sqlclient.SqlConnection;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
