+++ vertx-sql-client/src/test/java/io/vertx/sqlclient/tck/DataTypeTestBase.java
@@ -1,9 +1,10 @@
-package io.vertx.sqlclient;
+package io.vertx.sqlclient.tck;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.ext.unit.TestContext;
+import io.vertx.sqlclient.SqlConnection;
 import org.junit.After;
 import org.junit.Before;
