+++ vertx-sql-client/src/test/java/io/vertx/sqlclient/tck/SimpleQueryTestBase.java
@@ -14,13 +14,16 @@
  * limitations under the License.
  *
  */
-package io.vertx.sqlclient;
+package io.vertx.sqlclient.tck;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
+import io.vertx.sqlclient.SqlClient;
+import io.vertx.sqlclient.SqlConnection;
+import io.vertx.sqlclient.Tuple;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
